hack to make portable on 32/64 bit systems

+++ test/speedtest_msg.c
@@ -117,7 +117,7 @@ int main(void)
   }
   t1 = clock();
   s1 = (double)(t1 - t0)*(1.0 / (double)CLOCKS_PER_SEC);
-  printf("%s\t%8.0f\t%lu\n", "JSON", imax/s1, len);
+  printf("%s\t%8.0f\t%u\n", "JSON", imax/s1, (unsigned) len);
   fflush(stdout);
 
   t0 = clock();
